fix: preserve float precision in SimHash weight calculation
- Use float array accumulator instead of int array to maintain full precision
- Avoids truncation of vector weights during SimHash generation
- Improves quality of vector sketches for future codebook retraining

Addresses PR review comment about precision loss

+++ src/main/java/io/github/panghy/vectorsearch/FdbVectorSearch.java
@@ -750,11 +750,11 @@ public class FdbVectorSearch implements VectorSearch {
       byte[] counterKey = metaSubspace.pack(Tuple.from("node_id_counter"));
 
       return tr.get(counterKey).thenCompose(counterBytes -> {
-        long startId = counterBytes == null ? 1L : (long) decodeInt(counterBytes) + 1;
+        long startId = counterBytes == null ? 1L : ((long) decodeInt(counterBytes)) + 1;
         long endId = startId + vectors.size() - 1;
 
         // Update counter for next batch
-        tr.set(counterKey, encodeInt((int) endId));
+        tr.set(counterKey, encodeInt(endId));
 
         // Create ID to vector mapping
         Map<Long, float[]> idToVector = new LinkedHashMap<>();
